requirements.txt
Flask==2.3.3
Flask-WTF==1.2.1
email-validator==2.1.0.post1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
pytest==7.4.3
python-dotenv==1.0.0
Flask-Login==0.6.3
Pillow==10.1.0
pytest-cov==4.1.0
fastai==2.7.13
